Successful PoC on Android
- Removed demo code for the reat-native-audio-streaming library, since it
  doesn't support the media formats we need.
- Added a stop button that can make testing faster.

Often I'll see these messages in `adb logcat`:

    W/MediaPlayer-JNI: MediaPlayer finalized without being released

The react-native-sound library doesn't report this error, and the log
contains no other useful information. So, this library is probably not a
good choice,  since it seems unreliable (and doesn't actually do
streaming), but it at least proves that this sort of thing is possible
cross-platform.